c/core: Fix possible clean-up problem in init_dataset()
If there is a memory allocation failure when copying the column keys,
the dynamically allocated `dataset->columns[]` array is left in an
inconsistent state because remaining elements are left in an
uninitialized state.  If the dataset is later cleaned up with
`pb_release()`, it will try and free memory referenced by these
uninitialized `dataset->columns[]` elements leading to undefined
behavior.  Fix it by setting the remaining `dataset->columns[]` elements
to `NULL` on failure.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>