c/core: Fix memory leak and stale pointers in add_property_to_set()
`realloc()` is called twice to reallocate two blocks of memory to a
larger size before checking the result of either.  The function returns
an error if either reallocation failed, but not before updating the
original pointer for the one that succeeded (if any).  This would result
in undefined behavior when the property set is cleaned up by
`pb_release()`.  Fix it by updating the original pointers for each
successful reallocation.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>