c/core: Allow 0 rows or 0 columns in init_dataset()
If the `num_of_columns` or `num_of_rows` parameter value is 0, some of
the memory allocations will have size 0, so the calls to `malloc()` may
return `NULL`, causing `init_dataset()` to return an error.  Change the
function to only allocate `dataset->columns` and `dataset->types` if
`num_of_columns` is not 0, and to only allocate `dataset->rows` if
`num_of_rows` is not 0, setting the pointers to `NULL` without returning
an error when not allocated.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>